Implemented manual input handling in the win32 console in preparation for autocomplete
Keys are received and printed manually now.
This kind of sucks because of the need to manually process space, backspace, etc.
But it was the only way I could figure out to intercept tabs and have them behave correctly.

When tab is pressed, it just pops up a message box for now.
Actual tab-to-autocomplete will be implemented separately.  This is the framework for it.

refs #15